Silence warnings about throw() deprecation
Silence narrow cast warnings in swig modules

Fix a bunch more narrowing warnings

+++ xbmc/addons/kodi-addon-dev-kit/include/kodi/addon-instance/Visualization.h
@@ -745,7 +745,7 @@ namespace addon
           addon->toAddon.addonInstance->m_instanceData->toKodi.transfer_preset(addon->toKodi.kodiInstance, it.c_str());
       }
 
-      return presets.size();
+      return static_cast<unsigned int>(presets.size());
     }
 
     inline static int ADDON_GetActivePreset(const AddonInstance_Visualization* addon)
